Ending up refactored changes
