+++ .env.example
@@ -1,4 +1,13 @@
-NODE_ENV=development
+# Enviroment
+NODE_ENV=node_env
+
+# GraphQL settings
+GRAPHQL_PORT=port
+
+# MongoDB
 MONGO_USER=user
-MONGO_PASSWORD=password
+MONGO_PASSWORD=mongo@password
 MONGO_DB=mongodb
+
+# Security
+JWT_SECRET=jwt_secret
